+++ .travis.meta.yml
@@ -9,6 +9,8 @@ env:
   - GHCVER=7.10.2 CABALVER=1.22
   - GHCVER=7.10.2 CABALVER=1.22 STACKAGESNAPSHOT=lts-3.6
   - GHCVER=7.10.3 CABALVER=1.22
+  - GHCVER=7.10.3 CABALVER=1.22 STACKAGESNAPSHOT=lts-4.2
+  - GHCVER=7.10.3 CABALVER=1.22 STACKAGESNAPSHOT=lts-5.0
   - GHCVER=head CABALVER=1.22
 
 matrix:
