+++ aeson-extra.cabal
@@ -31,9 +31,9 @@ library
   ghc-options: -Wall
   build-depends:
       base                     >=4.6  && <4.9
-    , base-compat              >=0.6.0   && <0.10
     , aeson                    >=0.7.0.6 && <0.11
     , aeson-compat             >=0.3.0.0 && <0.4
+    , base-compat              >=0.6.0   && <0.10
     , bytestring               >=0.10 && <0.11
     , containers               >=0.5  && <0.6
     , exceptions               >=0.8  && <0.9
@@ -48,6 +48,15 @@ library
     , vector                   >=0.10 && <0.12
   exposed-modules:
       Data.Aeson.Extra
+      Data.Aeson.Extra.CollapsedList
+      Data.Aeson.Extra.Map
+      Data.Aeson.Extra.Time
+
+  if impl(ghc >= 7.8)
+    exposed-modules:
+      Data.Aeson.Extra.SingObject
+      Data.Aeson.Extra.SymTag
+
   default-language: Haskell2010
 
 test-suite aeson-extra-test
@@ -58,9 +67,9 @@ test-suite aeson-extra-test
   ghc-options: -Wall
   build-depends:
       base                     >=4.6  && <4.9
-    , base-compat              >=0.6.0   && <0.10
     , aeson                    >=0.7.0.6 && <0.11
     , aeson-compat             >=0.3.0.0 && <0.4
+    , base-compat              >=0.6.0   && <0.10
     , bytestring               >=0.10 && <0.11
     , containers               >=0.5  && <0.6
     , exceptions               >=0.8  && <0.9
