+++ package.yaml
@@ -16,9 +16,9 @@ ghc-options: -Wall
 
 dependencies:
   - base                     >=4.6  && <4.9
-  - base-compat              >=0.6.0   && <0.10
   - aeson                    >=0.7.0.6 && <0.11
   - aeson-compat             >=0.3.0.0 && <0.4
+  - base-compat              >=0.6.0   && <0.10
   - bytestring               >=0.10 && <0.11
   - containers               >=0.5  && <0.6
   - exceptions               >=0.8  && <0.9
@@ -31,7 +31,6 @@ dependencies:
   - time-parsers             >=0.1.0.0 && <0.2
   - unordered-containers     >=0.2  && <0.3
   - vector                   >=0.10 && <0.12
-
 library:
   source-dirs: src
