+++ stack-lts-4.yaml
@@ -1,0 +1,4 @@
+flags: {}
+packages:
+- '.'
+resolver: lts-4.2
