+++ stack-lts-5.yaml
@@ -1,0 +1,4 @@
+flags: {}
+packages:
+- '.'
+resolver: lts-5.0
